check if the user trying to delete the note his the one who created it

+++ backend/controllers/note/delete_note.js
@@ -1,14 +1,22 @@
 const Note = require('../../models/Note_model');
-const mongoose = require('mongoose');
 
 const deleteNote = async (req, res) => {
   const { noteID } = req.params;
-  console.log(req.params.noteID);
+  const userID = req.userID;
+
   try {
-    const note = await Note.findByIdAndDelete(noteID);
+    const note = await Note.findById(noteID);
     if (!note) {
       return res.status(404).json({ success: false, message: 'Note not found' });
     }
+
+    if (note.user.toString() !== userID) {
+      return res
+        .status(403)
+        .json({ success: false, message: 'You do not have permission to delete this note' });
+    }
+
+    await Note.findByIdAndDelete(noteID);
 
     return res.status(200).json({ success: true, message: 'Note deleted successfully' });
   } catch (error) {
